+++ src/bos/operators/power_on.py
@@ -26,6 +26,7 @@ from collections import defaultdict
 import logging
 from requests import HTTPError
 
+from bos.common.utils import exc_type_msg
 from bos.common.values import Action, Status
 import bos.operators.utils.clients.bss as bss
 import bos.operators.utils.clients.pcs as pcs
@@ -59,6 +60,8 @@ class PowerOnOperator(BaseOperator):
         ]
 
     def _act(self, components):
+        if not components:
+            return components
         self._preset_last_action(components)
         try:
             self._set_bss(components)
@@ -85,6 +88,10 @@ class PowerOnOperator(BaseOperator):
         Because the connection to the BSS tokens database can be lost due to
         infrequent use, retry up to retries number of times.
         """
+        if not components:
+            # If we have been passed an empty list, there is nothing to do.
+            LOGGER.debug("_set_bss: No components to act on")
+            return
         parameters = defaultdict(set)
         sessions = {}
         for component in components:
@@ -106,9 +113,9 @@ class PowerOnOperator(BaseOperator):
                 resp = bss.set_bss(node_set=nodes, kernel_params=kernel_parameters,
                                    kernel=kernel, initrd=initrd)
                 resp.raise_for_status()
-            except HTTPError:
-                LOGGER.error(f"Failed to set BSS for boot artifacts: {key} for"
-                             "nodes: {nodes}. Error: {err}")
+            except HTTPError as err:
+                LOGGER.error("Failed to set BSS for boot artifacts: %s for nodes: %s. Error: %s",
+                             key, nodes, exc_type_msg(err))
             else:
                 token = resp.headers['bss-referral-token']
                 attempts = 0
@@ -118,7 +125,8 @@ class PowerOnOperator(BaseOperator):
                         break
                     except Exception as err:
                         attempts += 1
-                        LOGGER.error(f"An error occurred attempting to record the BSS token: {err}")
+                        LOGGER.error("An error occurred attempting to record the BSS token: %s",
+                                     exc_type_msg(err))
                         if attempts > retries:
                             raise
                         LOGGER.info("Retrying to record the BSS token.")
@@ -128,6 +136,8 @@ class PowerOnOperator(BaseOperator):
                                        "desired_state": {"bss_token": token},
                                        "session": sessions[node]})
         LOGGER.info('Found %d components that require BSS token updates', len(bss_tokens))
+        if not bss_tokens:
+            return
         redacted_component_updates = [
             { "id": comp["id"], 
               "session": comp["session"]
@@ -135,7 +145,6 @@ class PowerOnOperator(BaseOperator):
             for comp in bss_tokens ]
         LOGGER.debug('Updated components (minus desired_state data): {}'.format(redacted_component_updates))
         self.bos_client.components.update_components(bss_tokens)
-
 
 if __name__ == '__main__':
     main(PowerOnOperator)