+++ src/bos/operators/utils/clients/cfs.py
@@ -25,7 +25,7 @@ from collections import defaultdict
 import logging
 from requests.exceptions import HTTPError, ConnectionError
 
-from bos.common.utils import requests_retry_session, PROTOCOL
+from bos.common.utils import compact_response_text, exc_type_msg, requests_retry_session, PROTOCOL
 
 SERVICE_NAME = 'cray-cfs-api'
 BASE_ENDPOINT = "%s://%s/v3" % (PROTOCOL, SERVICE_NAME)
@@ -51,11 +51,11 @@ def get_components(session=None, **params):
         LOGGER.debug("GET %s with params=%s", COMPONENTS_ENDPOINT, params)
         response = session.get(COMPONENTS_ENDPOINT, params=params)
         LOGGER.debug("Response status code=%d, reason=%s, body=%s", response.status_code,
-                     response.reason, response.text)
+                     response.reason, compact_response_text(response.text))
         try:
             response.raise_for_status()
         except HTTPError as err:
-            LOGGER.error("Failed getting nodes from CFS: %s", err)
+            LOGGER.error("Failed getting nodes from CFS: %s", exc_type_msg(err))
             raise
         response_json = response.json()
         new_components = response_json["components"]
@@ -75,11 +75,11 @@ def patch_components(data, session=None):
     LOGGER.debug("PATCH %s with body=%s", COMPONENTS_ENDPOINT, data)
     response = session.patch(COMPONENTS_ENDPOINT, json=data)
     LOGGER.debug("Response status code=%d, reason=%s, body=%s", response.status_code,
-                 response.reason, response.text)
+                 response.reason, compact_response_text(response.text))
     try:
         response.raise_for_status()
     except HTTPError as err:
-        LOGGER.error("Failed asking CFS to configure nodes: %s", err)
+        LOGGER.error("Failed asking CFS to configure nodes: %s", exc_type_msg(err))
         raise